rename to muacrypt
