+++ core/README.rst
@@ -1,13 +1,20 @@
 
-muacrypt: utilities for encrypting mail agents
-==============================================
+muacrypt: Autocrypt and more for mail user agents
+=================================================
 
 **this tool is under heavy development, use at your own risk**
 
-The muacrypt package offers a command line tool and API.
+``muacrypt`` helps mail user agents to process incoming and outgoing
+mails and keep Autocrypt state according to the `Autocrypt Level 1
+specification <https://autocrypt.org/autocrypt-spec-1.0.0.pdf>`_.
+``muacrypt`` manages accounts which maintain Autocrypt related state.
+Each account processes incoming and outgoing mail, parsing and adding
+Autocrypt headers as appropriate.  Functionality is exposed through
+a command line tool ``muacrypt`` and a Python api ``import muacrypt``.
+
 This README is intended to help contributors to get setup with running
 tests and using the command line. The online docs at
-https://py-autocrypt.readthedocs.io contain
+https://muacrypt.readthedocs.io contain
 more documentation about overall goals of the project.
 
 testing
@@ -27,12 +34,15 @@ To use the code and run tests you need to have installed:
 - tox either installed via ``pip install tox``
   or ``apt install python-tox``.
 
-If all requirements are fullfilled, simply issue::
+In one command on Debian::
+
+    apt install gnupg2 rng-tools python2.7-dev python3.5-dev python-tox
+
+Afterwards you can run all tests:
 
     $ tox
 
-to run all the tests against the autocrypt classes
-and the command line.
+this runs all automated tests.
 
 
 installation
@@ -41,13 +51,13 @@ installation
 You'll need the command line client "gpg", optionally "gpg2",
 available through "gnupg" and "gnugp2" on debian.
 
-To install the autocrypt command line tool you can install
-the "autocrypt" python package into your virtual environment
+To install the muacrypt command line tool you can install
+the "muacrypt" python package into your virtual environment
 of choice.  If you don't know about python's virtual environments
 you may just install the debian package "python-pip" and then
-use "pip" to install the autocrypt library and command line too::
+use "pip" to install the muacrypt library and command line too::
 
-    $ sudo pip install autocrypt
+    $ sudo pip install muacrypt
 
 This will install the required dependency "click", a python
 framework for writing command line clients.
@@ -66,7 +76,7 @@ and activate a python virtualenv and then once use
     $ pip install -e .
 
 Changes you subsequently make to the sources will be
-available without further installing the autocrypt
+available without further installing the muacrypt
 package again.
 
 
@@ -75,26 +85,26 @@ running the command line
 
 After installation simply run the main command::
 
-    autocrypt
+    muacrypt
 
 to see available sub commands and options.  Start by
 initializing an Autocrypt account which will maintain
 its own keyring and not interfere with your possibly
 existing gpg default keyring::
 
-    $ autocrypt init
+    $ muacrypt init
 
-Afterwards you can create an autocrypt header
+Afterwards you can create an Autocrypt header
 for an email address::
 
-    $ autocrypt make-header x@example.org
+    $ muacrypt make-header x@example.org
 
 You can process and integrate peer's Autocrypt
 keys by piping an email message into the ``process-incoming`` subcommand::
 
-    $ autocrypt process-incoming <EMAIL_MESSAGE_FILE
+    $ muacrypt process-incoming <EMAIL_MESSAGE_FILE
 
-At any point you can show the status of your autocrypt
+At any point you can show the status of your muacrypt
 account::
 
-    $ autocrypt show
+    $ muacrypt status
