--- core/experiments/README.md
@@ -1,1 +1,0 @@
-Functions to generate and manipulate GPG keys with PGPy library.
