--- core/experiments/_version.py
@@ -1,1 +1,0 @@
-version = '0.1.0'
