--- core/experiments/requirements.txt
@@ -1,1 +1,0 @@
-PGPy>=0.4.0
