+++ core/muacrypt/cmdline.py
@@ -22,9 +22,9 @@ from .bot import bot_reply
 
 @click.command(cls=MyGroup, context_settings=dict(help_option_names=["-h", "--help"]))
 @click.option("--basedir", type=click.Path(),
-              default=click.get_app_dir("autocrypt"),
-              envvar="AUTOCRYPT_BASEDIR",
-              help="directory where autocrypt account state is stored")
+              default=click.get_app_dir("muacrypt"),
+              envvar="MUACRYPT_BASEDIR",
+              help="directory where muacrypt state is stored")
 @click.version_option()
 @click.pass_context
 def autocrypt_main(context, basedir):
@@ -35,12 +35,12 @@ def autocrypt_main(context, basedir):
 
 @mycommand()
 @click.option("--replace", default=False, is_flag=True,
-              help="delete autocrypt account directory before attempting init")
+              help="delete muacrypt state directory before attempting init")
 @click.option("--no-identity", default=False, is_flag=True,
               help="initializing without creating a default identity")
 @click.pass_context
 def init(ctx, replace, no_identity):
-    """init autocrypt account state.
+    """init muacrypt state.
 
     By default this command creates account state in a directory with
     a default "catch-all" identity which matches all email addresses
@@ -112,7 +112,7 @@ def add_identity(ctx, identity_name, use_system_keyring,
     existing key with --use-key=keyhandle where keyhandle may be
     something for which gpg finds it with 'gpg --list-secret-keys keyhandle'.
     Typically you will then also specify --use-system-keyring to make use of
-    your existing keys.  All incoming autocrypt keys will thus be stored in
+    your existing keys.  All incoming muacrypt keys will thus be stored in
     the system key ring instead of an own keyring.
     """
     account = get_account(ctx)
@@ -179,7 +179,7 @@ def test_email(ctx, emailadr):
 @click.argument("emailadr", type=click.STRING)
 @click.pass_context
 def make_header(ctx, emailadr):
-    """print autocrypt header for an emailadr. """
+    """print Autocrypt header for an emailadr. """
     account = get_account(ctx)
     click.echo(account.make_header(emailadr))
 
@@ -202,7 +202,7 @@ def set_prefer_encrypt(ctx, value):
 @mycommand("process-incoming")
 @click.pass_context
 def process_incoming(ctx):
-    """parse autocrypt headers from stdin mail. """
+    """parse Autocrypt headers from stdin mail. """
     account = get_account(ctx)
     msg = mime.parse_message_from_file(sys.stdin)
     r = account.process_incoming(msg)
@@ -217,7 +217,7 @@ def process_incoming(ctx):
 @mycommand("process-outgoing")
 @click.pass_context
 def process_outgoing(ctx):
-    """add autocrypt header for outgoing mail.
+    """add Autocrypt header for outgoing mail.
 
     We process mail from stdin by adding an Autocrypt
     header and send the resulting message to stdout.
@@ -289,7 +289,7 @@ def export_public_key(ctx, id, keyhandle_or_email):
 @id_option
 @click.pass_context
 def export_secret_key(ctx, id):
-    """print secret key of own autocrypt account. """
+    """print secret key of own account."""
     account = get_account(ctx)
     ident = account.get_identity(id)
     data = ident.export_secret_key()
@@ -299,7 +299,7 @@ def export_secret_key(ctx, id):
 @mycommand()
 @click.pass_context
 def status(ctx):
-    """print account and identity info. """
+    """print account info and status. """
     account = get_account(ctx)
     _status(account)
