+++ core/setup.py
@@ -2,7 +2,7 @@ import os
 from setuptools import setup
 
 def main():
-    with open(os.path.join("autocrypt", "__init__.py")) as f:
+    with open(os.path.join("muacrypt", "__init__.py")) as f:
         for line in f:
             if "__version__" in line.strip():
                 version = line.split("=", 1)[1].strip().strip('"')
@@ -12,14 +12,14 @@ def main():
         long_desc = f.read()
 
     setup(
-        name='autocrypt',
+        name='muacrypt',
         description='Autocrypt: E-mail Encryption for Everyone example implementation',
         long_description = long_desc,
         version=version,
-        url='https://autocrypt.org',
+        url='https://muacrypt.readthedocs.io',
         license='MIT license',
         platforms=['unix', 'linux', 'osx', 'cygwin', 'win32'],
-        author='holger krekel and the autocrypt team',
+        author='holger krekel and the muacrypt team',
         author_email='holger@merlinux.eu',
         classifiers=['Development Status :: 3 - Alpha',
                      'Intended Audience :: Developers',
@@ -29,10 +29,10 @@ def main():
                      'Topic :: Utilities',
                      'Intended Audience :: Developers',
                      'Programming Language :: Python'],
-        packages=['autocrypt'],
+        packages=['muacrypt'],
         entry_points='''
             [console_scripts]
-            autocrypt=autocrypt.cmdline:autocrypt_main
+            muacrypt=muacrypt.cmdline:autocrypt_main
         ''',
         install_requires = ["click>=6.0", "six", "attrs"],
         zip_safe=False,
